Assignment Part 1 of 3
This is an assignment to scrape a wiki page describing the postcodes, boroughs & neighbourhoods in the city of Toronto, invoke an API to add geo location data, and then segment or cluster the data based on interest (keywords) & visualize the clusters. The significant python libraries that will be used to accomplish this assignment include pandas, requests, BeautifulSoup, & Folium for visualization.