Assignment part 3 of 3
In part 3 of the assignment, I've used FourSquare API to retrieve the venues near each longitude & latitude, regrouped by neighbourhoods, and then by top 5 categories of venues in each neighbourhood. All the venue categories are transformed into truth values using one-hot-encoding which allows us to measure the mean frequency of occurrence of each category. The data is then fed to a K-Nearest Neighbours algorithm to supervise and generate 5 types of clusters. The exercise shows that all the parks and recreational areas are clustered (#2 - third) together & one type of cluster (#0 - first) is the most prevalent of all.