Capstone Project Report - Intro & Data sections
The report utilizes IBM Data Science methodology starting with the Introduction (Business Problem) section, immediately followed by methodology & data section. The report is a baseline and will be completed along with the project. Separately, the Jupyter notebook will be made available in this repository, along with the URLs to the data utilized for this project. For questions, write to mbehara@AdelleTech.com.