Classify Numerical Series Data using SciKit Learn
This notebook demonstrates how to classify numerical series data such as the data from a HBV (Heart Beat Variability) piezo-electric sensor. Useful reference for crating a procedure to recursively pick up all the comma delimited files into a pandas data frame & then use a NumPy array to turn around and use the histogram feature to bin the data (feature engineering). Sine all the features are in numerical form and are just counts, there was no need for further normalization. This notebook is a good reference for data wrangling using pandas and numpy and straight up processing of lists and tuples. This model is a prototype that needs expert evaluation & fine tuning for consistency before deploying in production. The Logistic Regression model was able to classify the data with a max 75% accuracy & a Support Vector Machine(SVM) was able to do so with max 65% accuracy. Never ceases to amaze me how sophisticated scikit learn library is in it's ability to quickly encode class variables in a pandas dataframe.